+++ GetThePicture/Copybook/Compiler/Storage/Base/IStorageNode.cs
@@ -1,5 +1,3 @@
-using GetThePicture.Copybook.Compiler.Layout;
-
 namespace GetThePicture.Copybook.Compiler.Storage.Base;
 
 public interface IStorageNode