+++ GetThePicture/Picture/Clause/Codec/Semantic/Time/Decoder.cs
@@ -10,11 +10,8 @@ internal static class Decoder
 {
     public static TimeOnly Decode(ReadOnlySpan<byte> buffer, PicMeta pic)
     {
-        if (pic.BaseClass == PicBaseClass.Numeric && pic.Signed)
-            throw new NotSupportedException($"Unsupported TimeOnly base type: PIC S9");
-
-        if (pic.Usage != PicUsage.Display)
-            throw new NotSupportedException($"'Time' does not support usage '{pic.Usage}'. Only DISPLAY is allowed.");
+        Constraint rule = Rules.GetConstraint(pic.Semantic);
+        rule.ValidateOrThrow(pic, pic.Semantic.ToString());
 
         // TODO: 看看要不要支援 COMP-3 (PACKED-DECIMAL)