+++ GetThePicture/Picture/Clause/Codec/Semantic/Time/Encoder.cs
@@ -11,8 +11,8 @@ internal static class Encoder
         if (value is not TimeOnly time)
             throw new FormatException($"Invalid value type for TimeOnly encoding: {value.GetType().FullName}");
 
-        if (pic.Usage != PicUsage.Display)
-            throw new NotSupportedException($"'Time' does not support usage '{pic.Usage}'. Only DISPLAY is allowed.");
+        Constraint rule = Rules.GetConstraint(pic.Semantic);
+        rule.ValidateOrThrow(pic, pic.Semantic.ToString());
 
         string s = pic.Semantic switch
         {